Create task-1.java.codsoft
Project Description: "Guess the Number Challenge"

Overview:
The "Guess the Number Challenge" is a fun and interactive game designed to test your guessing skills. In this project, we've created a simple yet engaging game that involves guessing a random number within a specified range. The game incorporates various features to make it more exciting and enjoyable.
How to Play:

Start the game and choose a difficulty level (specify the range and number of attempts).

The game will generate a random number within the chosen range.

Enter your first guess.

Receive feedback on whether your guess is correct, too high, or too low.

Continue guessing within the specified number of attempts until you guess the correct number.

If you want to play another round, choose to restart and improve your score.

Objective:
The objective of the "Guess the Number Challenge" is to have fun while honing your guessing skills. Try to guess the correct number within the allotted attempts and challenge yourself to achieve the highest score possible.

So, are you ready to take on the "Guess the Number Challenge" and see how well you can guess random numbers? Good luck, and may the best guesser win!